Added controler actions

+++ src/main/java/hr/fer/marin/zavrsni/repository/ObjectRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ObjectRepository extends JpaRepository<Object, Long> {
+public interface ObjectRepository extends JpaRepository<Object, Integer> {
 }
